Add README file with instructions for installing and using the program
This contribution adds a README file to the project, which describes how to install and use the Email Verifier program. Prerequisites, installation, usage, examples, and contributions are all covered in the README. Users should be able to learn how to use the program and contribute to the project more easily as a result of this.